+++ Cargo.toml
@@ -99,6 +99,7 @@ ctor = "0.2.9"
 dashmap = "6.0.1"
 datafusion = { path = "datafusion/core", version = "45.0.0", default-features = false }
 datafusion-catalog = { path = "datafusion/catalog", version = "45.0.0" }
+datafusion-catalog-datasource = { path = "datafusion/catalog-datasource", version = "45.0.0", default-features = false }
 datafusion-catalog-listing = { path = "datafusion/catalog-listing", version = "45.0.0", default-features = false }
 datafusion-common = { path = "datafusion/common", version = "45.0.0", default-features = false }
 datafusion-common-runtime = { path = "datafusion/common-runtime", version = "45.0.0" }
