+++ datafusion/catalog-datasource/LICENSE.txt
@@ -1,0 +1,1 @@
+../../LICENSE.txt
