+++ datafusion/catalog-datasource/src/write/demux.rs
