+++ datafusion/catalog-listing/src/helpers.rs
@@ -20,9 +20,9 @@
 use std::mem;
 use std::sync::Arc;
 
-use super::ListingTableUrl;
-use super::PartitionedFile;
 use datafusion_catalog::Session;
+use datafusion_catalog_datasource::ListingTableUrl;
+use datafusion_catalog_datasource::PartitionedFile;
 use datafusion_common::internal_err;
 use datafusion_common::{HashMap, Result, ScalarValue};
 use datafusion_expr::{BinaryExpr, Operator};
