+++ datafusion/core/src/datasource/physical_plan/file_stream.rs
@@ -37,8 +37,10 @@ use crate::physical_plan::RecordBatchStream;
 use arrow::datatypes::SchemaRef;
 use arrow::error::ArrowError;
 use arrow::record_batch::RecordBatch;
-pub use datafusion_catalog_listing::file_stream::{FileOpenFuture, FileOpener, OnError};
-use datafusion_catalog_listing::file_stream::{
+pub use datafusion_catalog_datasource::file_stream::{
+    FileOpenFuture, FileOpener, OnError,
+};
+use datafusion_catalog_datasource::file_stream::{
     FileStreamMetrics, FileStreamState, NextOpen,
 };
 use datafusion_common::ScalarValue;
